[gl] Moved TexturePool to gl namespace
