+++ src/mbgl/util/raster.hpp
@@ -2,7 +2,7 @@
 #define MBGL_UTIL_RASTER
 
 #include <mbgl/gl/gl.hpp>
-#include <mbgl/util/texture_pool.hpp>
+#include <mbgl/gl/texture_pool.hpp>
 #include <mbgl/util/image.hpp>
 #include <mbgl/util/ptr.hpp>
 #include <mbgl/util/chrono.hpp>
@@ -14,7 +14,7 @@ namespace mbgl {
 class Raster : public std::enable_shared_from_this<Raster> {
 
 public:
-    Raster(TexturePool&);
+    Raster(gl::TexturePool&);
     ~Raster();
 
     // load image data
@@ -50,7 +50,7 @@ private:
     bool loaded = false;
 
     // shared texture pool
-    TexturePool& texturePool;
+    gl::TexturePool& texturePool;
 
     // min/mag filter
     GLint filter = 0;
